requirements.txt
# HTTP client for making API requests to Omada and Azure
httpx>=0.28.0

# Environment variable loading from .env files
python-dotenv>=1.1.0

# Model Context Protocol (MCP) server framework
mcp>=1.13.0

# FastMCP - High-level MCP server framework
fastmcp>=2.12.0

# Testing framework (for VS Code test integration)
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Requests library (used in some test files)
requests>=2.31.0